updated breadcrumbs and navbar to be client rendered and layout.js to be server rendered
