+++ silverpi/src/app/home/layout.js
@@ -1,48 +1,11 @@
-'use client'
 import Navbar from "../components/navbar.js"
-import { useRouter, usePathname, useSearchParams} from "next/navigation";
-import { useEffect, useState } from "react";
-import Breadcrumb from "../components/breadcrumbs.js";
-import BreadcrumbItem from "../components/breadcrumbitem.js";
+import BreadcrumbGenerator from "../components/breadcrumbgenerator.js";
 
 export default function DashboardLayout({ children }) {
-    const [breadcrumbs, setBreadcrumbs] = useState();
-    var pathName = usePathname()
-
-    useEffect(() => {
-      let pathArray =  pathName.split("/");
-      if(pathArray.length <= 2) {
-        var breadcrumbs = []
-        setBreadcrumbs(breadcrumbs);
-      } else {
-        pathArray.shift();
-
-        pathArray = pathArray.filter((path) => path !== "");
-
-        const breadcrumbs = pathArray.map((path, index) => {
-        const href = "/" + pathArray.slice(0, index + 1).join("/");
-        return {
-            href,
-            label: path.charAt(0).toUpperCase() + path.slice(1),
-        };
-        });
-        setBreadcrumbs(breadcrumbs);
-      }
-      
-    }, [pathName]);
-
     return (
         <>
             <Navbar />
-            <Breadcrumb>
-                {
-                breadcrumbs &&
-                breadcrumbs.map((breadcrumb) => (
-                    <BreadcrumbItem key={breadcrumb.href} href={breadcrumb.href}>
-                    <span className = "text-black hover:underline">{breadcrumb.label}</span>
-                    </BreadcrumbItem>
-                ))}
-            </Breadcrumb>
+            <BreadcrumbGenerator></BreadcrumbGenerator>
             <section>{children}</section>
         </>
     );
